Error: socket hang up

+++ models/comment.go
@@ -2,7 +2,6 @@ package models
 
 import (
 	"time"
-
 	"gorm.io/gorm"
 )
 
@@ -10,7 +9,7 @@ type Comment struct {
 	ID           uint       `gorm:"primaryKey"`
 	DateTime     time.Time  `json:"dateTime"`
 	MyComment    string     `json:"myComment"`
-	Rating       float32    `json:"rating"`
+	Rating       float64    `json:"rating"`
 	UserID       uint       `json:"userId"`
 	User         User       `gorm:"foreignKey:UserID" json:"user"`
 	RestaurantID uint       `json:"restaurantId"`
